Divided file : Enum GameType , NumberOfPlayers , Message in InputProperty
